src/test/javascript/spec/app/entities/delivery/delivery.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import * as moment from 'moment';
import { DATE_FORMAT } from 'app/shared/constants/input.constants';
import { DeliveryService } from 'app/entities/delivery/delivery.service';
import { IDelivery, Delivery } from 'app/shared/model/delivery.model';

describe('Service Tests', () => {
  describe('Delivery Service', () => {
    let injector: TestBed;
    let service: DeliveryService;
    let httpMock: HttpTestingController;
    let elemDefault: IDelivery;
    let expectedResult: IDelivery | IDelivery[] | boolean | null;
    let currentDate: moment.Moment;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(DeliveryService);
      httpMock = injector.get(HttpTestingController);
      currentDate = moment();

      elemDefault = new Delivery(0, 0, 'AAAAAAA', currentDate, 0, 0, 0);
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign(
          {
            deliveryDate: currentDate.format(DATE_FORMAT)
          },
          elemDefault
        );

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a Delivery', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
            deliveryDate: currentDate.format(DATE_FORMAT)
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            deliveryDate: currentDate
          },
          returnedFromService
        );

        service.create(new Delivery()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a Delivery', () => {
        const returnedFromService = Object.assign(
          {
            idDelivery: 1,
            deliveryDescription: 'BBBBBB',
            deliveryDate: currentDate.format(DATE_FORMAT),
            subTotal: 1,
            vatamount: 1,
            total: 1
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            deliveryDate: currentDate
          },
          returnedFromService
        );

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of Delivery', () => {
        const returnedFromService = Object.assign(
          {
            idDelivery: 1,
            deliveryDescription: 'BBBBBB',
            deliveryDate: currentDate.format(DATE_FORMAT),
            subTotal: 1,
            vatamount: 1,
            total: 1
          },
          elemDefault
        );

        const expected = Object.assign(
          {
            deliveryDate: currentDate
          },
          returnedFromService
        );

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a Delivery', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
